partitions: hack to force active part based on environment
Currently there is code that tries to determain which is the
active partition based on what is mounted.

On one of the systems that we are using the root mount point
looks like this:

    /dev/root /rom

The Mender client is not able to figure this out and simply
errors out saying it is not able to determain the active
partition.

This code is a hack that disables the "probing" and simply
uses the mender_boot_part environment variable to determain
which partition is active.

Signed-off-by: Mirza Krak <mirza.krak@northern.tech>